refactor: deep nest island in folders
