+++ islands/tools/expression-visualizer/ExpressionVisualizationList.tsx
@@ -3,9 +3,9 @@ import type { JSX } from "preact";
 //? State management
 import { useState } from "preact/hooks";
 //? Component
-import { ExpressionVisualizationListItem } from "../../components/tools/expression-visualizer/ExpressionVisualizationListItem.tsx";
+import { ExpressionVisualizationListItem } from "../../../components/tools/expression-visualizer/ExpressionVisualizationListItem.tsx";
 //? Type
-import type { visualizer } from "../../types/component-properties/tools/expression-visualizer/Visualizer.ts";
+import type { visualizer } from "../../../types/component-properties/tools/expression-visualizer/Visualizer.ts";
 
 //? Renders an Expression list
 export default function ExpressionVisualizationList(
