+++ islands/tools/expression-visualizer/ExpressionVisualizerAdvanced.tsx
@@ -1,14 +1,14 @@
 //? State management
 import { useState } from "preact/hooks";
 //? Components
-import { StyledButton } from "../../components/UI/StyledButton.tsx";
-import { StyledInput } from "../../components/UI/StyledInput.tsx";
+import { StyledButton } from "../../../components/UI/StyledButton.tsx";
+import { StyledInput } from "../../../components/UI/StyledInput.tsx";
 import ExpressionVisualizationList from "./ExpressionVisualizationList.tsx";
 //? Validation
-import { validationStatus } from "../../types/forms/validationStatus.ts";
-import { validateNonEmptyText } from "../../services/form-validation/validateNonEmptyText.ts";
+import { validationStatus } from "../../../types/forms/validationStatus.ts";
+import { validateNonEmptyText } from "../../../services/form-validation/validateNonEmptyText.ts";
 //? Type
-import type { visualizer } from "../../types/component-properties/tools/expression-visualizer/Visualizer.ts";
+import type { visualizer } from "../../../types/component-properties/tools/expression-visualizer/Visualizer.ts";
 type ValidationStatuses<K extends keyof visualizer> = Record<
   K,
   validationStatus
