+++ islands/tools/whatsapp-message-link-generator/WhatsappLinkGenerator.tsx
@@ -1,23 +1,23 @@
 //? Manage form state
 import { useEffect, useRef, useState } from "preact/hooks";
 //? Styled components
-import { StyledInput } from "../../components/UI/StyledInput.tsx";
-import { StyledButton } from "../../components/UI/StyledButton.tsx";
-import { CountryPhoneCodeSelect } from "../../components/UI/CountryPhoneCodeSelect.tsx";
+import { StyledInput } from "../../../components/UI/StyledInput.tsx";
+import { StyledButton } from "../../../components/UI/StyledButton.tsx";
+import { CountryPhoneCodeSelect } from "../../../components/UI/CountryPhoneCodeSelect.tsx";
 //? Types for typecasting
-import { validationStatus } from "../../types/forms/validationStatus.ts";
-import type { WhatsappLinkData } from "../../types/component-properties/tools/whatsapp-link-generator/whatsapp-link-data.ts";
-import { WhatsappMessagePreview } from "../../components/tools/whatsapp-message-link-generator/WhatsappMessagePreview.tsx";
+import { validationStatus } from "../../../types/forms/validationStatus.ts";
+import type { WhatsappLinkData } from "../../../types/component-properties/tools/whatsapp-link-generator/whatsapp-link-data.ts";
+import { WhatsappMessagePreview } from "../../../components/tools/whatsapp-message-link-generator/WhatsappMessagePreview.tsx";
 
 //? Validation functions
-import { patternValidation } from "../../services/form-validation/patternValidation.ts";
-import { validateAreaCode } from "../../services/form-validation/validateAreaCode.ts";
-import { validatePhoneNumber } from "../../services/form-validation/validatePhoneNumber.ts";
-import { WhatsappLinksList } from "../../components/tools/whatsapp-message-link-generator/WhatsappLinksList.tsx";
-import { StyledTextArea } from "../../components/UI/StyledTextArea.tsx";
+import { patternValidation } from "../../../services/form-validation/patternValidation.ts";
+import { validateAreaCode } from "../../../services/form-validation/validateAreaCode.ts";
+import { validatePhoneNumber } from "../../../services/form-validation/validatePhoneNumber.ts";
+import { WhatsappLinksList } from "../../../components/tools/whatsapp-message-link-generator/WhatsappLinksList.tsx";
+import { StyledTextArea } from "../../../components/UI/StyledTextArea.tsx";
 
 //? Defined variable to replace in message body
-import { WHATSAPP_MESSAGE_VARIABLE_PLACEHOLDER } from "../../data/tools/whatsapp-link-generator/whatsapp-message-variable-placeholder.ts";
+import { WHATSAPP_MESSAGE_VARIABLE_PLACEHOLDER } from "../../../data/tools/whatsapp-link-generator/whatsapp-message-variable-placeholder.ts";
 
 //? Default form values and validation
 const baseLinkData: WhatsappLinkData = {
