+++ routes/projects/food-order/index.tsx
@@ -7,7 +7,7 @@ import { Base } from "../../../components/base/Base.tsx";
 //? Navigation Buttons to go back to the previous page
 import { NavigationButtons } from "../../../components/misc/NavigationButtons.tsx";
 //? Component responsible for rendering the food list and modals
-import FoodOrder from "../../../islands/projects/FoodOrder.tsx";
+import FoodOrder from "../../../islands/projects/food-order/FoodOrder.tsx";
 //? Import Food type to typecast the data received
 import { Food } from "../../../types/component-properties/projects/food-order/Food.ts";
 //? Import middleware responsible for pulling food items
